fix float number for carrier

+++ app/controllers/shippings_controller.rb
@@ -2,7 +2,7 @@ class ShippingsController < ApplicationController
   before_action :authenticate_user!
 
   def index
-    
+
   end
 
   def order_confirmation
@@ -24,7 +24,7 @@ class ShippingsController < ApplicationController
       @totalprice += item.price.to_f * item.quantity.to_i
     end
     @carrier.price.to_i
-      @totalprice = @totalprice + @carrier.price.to_i
+      @totalprice = @totalprice + @carrier.price.to_f
   end
 
 end
